requirements.txt
# Core scraping dependencies
scrapy>=2.11.0
pandas>=2.0.0
fake-useragent>=1.4.0
python-dotenv>=1.0.0

# Dashboard dependencies
streamlit>=1.28.0
plotly>=5.17.0
sqlalchemy>=2.0.0

# Testing dependencies (optional)
pytest>=7.4.0
syrupy>=4.0.0
